add e2e tests to check if /rows have signed urls
