+++ .github/workflows/_e2e_tests.yml
@@ -55,6 +55,13 @@ jobs:
           COMMON_BLOCKED_DATASETS: "__DUMMY_DATASETS_SERVER_USER__/blocked-*"
           ADMIN_HF_ORGANIZATION: "valid_org"
           E2E_ADMIN_USER_TOKEN: "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"
+          # the containers need to access S3 and Cloudfront
+          S3_ACCESS_KEY_ID: ${{ secrets.S3_ACCESS_KEY_ID }}
+          S3_SECRET_ACCESS_KEY: ${{ secrets.S3_SECRET_ACCESS_KEY }}
+          CLOUDFRONT_KEY_PAIR_ID: "K3814DK2QUJ71H"
+          CLOUDFRONT_PRIVATE_KEY: ${{ secrets.CLOUDFRONT_PRIVATE_KEY }}
+          ASSETS_BASE_URL: "https://datasets-server-test.us.dev.moon.huggingface.tech/assets"
+          CACHED_ASSETS_BASE_URL: "https://datasets-server-test.us.dev.moon.huggingface.tech/cached-assets"
         run: docker compose -f docker-compose-datasets-server.yml up -d --wait --wait-timeout 20
         working-directory: ./tools
       - name: Install poetry
